Added Waste Map Optimization
Waste Map uses a guillotine algorithm to track the wasted
space inside each shelf and attempts to pack items into it.

A bug in the Guillotine class was found and corrected and unit-tests were changed to account for this. Guillotine needs to get its rotation rules adjusted to better match Jukka's article.