Feature: Maximal Rectangles Partial Implementation
This commit begins the implementation of the Maximal
Rectangles algorithm. I'm trying to build as much of
this algorithm using staticmethods so as to simplify
testing.

The overall algorithm is not implemented in this build
but about 50% of the static methods that will make up
the algorithm are working and have unit tests.